+++ AlphaFS/Filesystem/DirectoryInfo Class/DirectoryInfo.cs
@@ -132,7 +132,7 @@ namespace Alphaleonis.Win32.Filesystem
                   Refresh();
 
                var attrs = Win32AttributeData.dwFileAttributes;
-               return DataInitialised == 0 && File.HasValidAttributes(attrs) && File.IsDirectory(attrs);
+               return DataInitialised == 0 && File.IsDirectory(attrs);
             }
             catch
             {